pyproject.toml
[tool.poetry]
name = "saucedemo-yu"
version = "0.0.1"
description = "stady"
authors = ["Yuriyvas <yuriy.vassiliev2010@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10.2"
pytest = "^7.2.1"
selenium = "^4.8.0"
webdriver-manager = "^3.8.5"
black = "^23.1.0"
flake8 = "^6.0.0"
allure-pytest = "^2.12.0"
allure-python-commons = "^2.12.0"
requests = "^2.28.2"
pylint = "^2.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

